(improvement)(Headless) Refactor service package of Headless-Server module
