+++ chat/server/src/main/java/com/tencent/supersonic/chat/server/rest/ChatConfigController.java
@@ -9,7 +9,7 @@ import com.tencent.supersonic.chat.api.pojo.response.ChatConfigResp;
 import com.tencent.supersonic.chat.api.pojo.response.ChatConfigRichResp;
 import com.tencent.supersonic.chat.server.service.ConfigService;
 import com.tencent.supersonic.headless.api.pojo.response.ItemResp;
-import com.tencent.supersonic.headless.server.web.service.SchemaService;
+import com.tencent.supersonic.headless.server.facade.service.SemanticLayerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -32,7 +32,7 @@ public class ChatConfigController {
     private ConfigService configService;
 
     @Autowired
-    private SchemaService schemaService;
+    private SemanticLayerService semanticLayerService;
 
     @PostMapping
     public Long addChatConfig(@RequestBody ChatConfigBaseReq extendBaseCmd,
@@ -70,7 +70,7 @@ public class ChatConfigController {
 
     @GetMapping("/getDomainDataSetTree")
     public List<ItemResp> getDomainDataSetTree() {
-        return schemaService.getDomainDataSetTree();
+        return semanticLayerService.getDomainDataSetTree();
     }
 
 }
