+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/facade/rest/MetaDiscoveryApiController.java
@@ -3,7 +3,7 @@ package com.tencent.supersonic.headless.server.facade.rest;
 import com.tencent.supersonic.auth.api.authentication.pojo.User;
 import com.tencent.supersonic.auth.api.authentication.utils.UserHolder;
 import com.tencent.supersonic.headless.api.pojo.request.QueryMapReq;
-import com.tencent.supersonic.headless.server.facade.service.RetrieveService;
+import com.tencent.supersonic.headless.server.facade.service.ChatQueryService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -20,14 +20,14 @@ import javax.servlet.http.HttpServletResponse;
 public class MetaDiscoveryApiController {
 
     @Autowired
-    private RetrieveService metaDiscovery;
+    private ChatQueryService chatQueryService;
 
     @PostMapping("map")
     public Object map(@RequestBody QueryMapReq queryMapReq,
                       HttpServletRequest request, HttpServletResponse response) throws Exception {
         User user = UserHolder.findUser(request, response);
         queryMapReq.setUser(user);
-        return metaDiscovery.map(queryMapReq);
+        return chatQueryService.map(queryMapReq);
     }
 
 }
