+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/facade/service/RetrieveService.java
@@ -1,16 +1,12 @@
 package com.tencent.supersonic.headless.server.facade.service;
 
-import com.tencent.supersonic.headless.api.pojo.request.QueryMapReq;
 import com.tencent.supersonic.headless.api.pojo.request.QueryReq;
-import com.tencent.supersonic.headless.api.pojo.response.MapInfoResp;
 import com.tencent.supersonic.headless.api.pojo.response.SearchResult;
 
 import java.util.List;
 
 public interface RetrieveService {
 
-    MapInfoResp map(QueryMapReq queryMapReq);
-
-    List<SearchResult> search(QueryReq queryCtx);
+    List<SearchResult> retrieve(QueryReq queryCtx);
 
 }
