+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/facade/service/impl/RetrieveServiceImpl.java
@@ -3,21 +3,12 @@ package com.tencent.supersonic.headless.server.facade.service.impl;
 import com.google.common.collect.Lists;
 import com.tencent.supersonic.auth.api.authentication.pojo.User;
 import com.tencent.supersonic.common.pojo.enums.DictWordType;
-import com.tencent.supersonic.common.pojo.enums.TimeDimensionEnum;
 import com.tencent.supersonic.headless.api.pojo.SchemaElement;
-import com.tencent.supersonic.headless.api.pojo.SchemaElementMatch;
 import com.tencent.supersonic.headless.api.pojo.SchemaElementType;
-import com.tencent.supersonic.headless.api.pojo.SchemaItem;
-import com.tencent.supersonic.headless.api.pojo.SchemaMapInfo;
 import com.tencent.supersonic.headless.api.pojo.SemanticSchema;
 import com.tencent.supersonic.headless.api.pojo.request.QueryFilter;
 import com.tencent.supersonic.headless.api.pojo.request.QueryFilters;
-import com.tencent.supersonic.headless.api.pojo.request.QueryMapReq;
 import com.tencent.supersonic.headless.api.pojo.request.QueryReq;
-import com.tencent.supersonic.headless.api.pojo.response.DataSetMapInfo;
-import com.tencent.supersonic.headless.api.pojo.response.DataSetResp;
-import com.tencent.supersonic.headless.api.pojo.response.MapInfoResp;
-import com.tencent.supersonic.headless.api.pojo.response.MapResp;
 import com.tencent.supersonic.headless.api.pojo.response.S2Term;
 import com.tencent.supersonic.headless.api.pojo.response.SearchResult;
 import com.tencent.supersonic.headless.chat.QueryContext;
@@ -25,18 +16,15 @@ import com.tencent.supersonic.headless.chat.knowledge.DataSetInfoStat;
 import com.tencent.supersonic.headless.chat.knowledge.DictWord;
 import com.tencent.supersonic.headless.chat.knowledge.HanlpMapResult;
 import com.tencent.supersonic.headless.chat.knowledge.KnowledgeBaseService;
-import com.tencent.supersonic.headless.chat.knowledge.builder.BaseWordBuilder;
 import com.tencent.supersonic.headless.chat.knowledge.helper.HanlpHelper;
 import com.tencent.supersonic.headless.chat.knowledge.helper.NatureHelper;
 import com.tencent.supersonic.headless.chat.mapper.MatchText;
 import com.tencent.supersonic.headless.chat.mapper.ModelWithSemanticType;
 import com.tencent.supersonic.headless.chat.mapper.SearchMatchStrategy;
-import com.tencent.supersonic.headless.server.pojo.MetaFilter;
-import com.tencent.supersonic.headless.server.web.service.ChatContextService;
-import com.tencent.supersonic.headless.server.facade.service.ChatQueryService;
-import com.tencent.supersonic.headless.server.web.service.DataSetService;
 import com.tencent.supersonic.headless.server.facade.service.RetrieveService;
-import com.tencent.supersonic.headless.server.web.service.SemanticLayerService;
+import com.tencent.supersonic.headless.server.web.service.ChatContextService;
+import com.tencent.supersonic.headless.server.web.service.DataSetService;
+import com.tencent.supersonic.headless.server.web.service.SchemaService;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.CollectionUtils;
 import org.springframework.beans.BeanUtils;
@@ -45,7 +33,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
@@ -54,7 +41,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 
 @Service
@@ -67,40 +53,22 @@ public class RetrieveServiceImpl implements RetrieveService {
     private DataSetService dataSetService;
 
     @Autowired
-    private ChatQueryService chatQueryService;
-
-    @Autowired
     private ChatContextService chatContextService;
 
     @Autowired
-    private SemanticLayerService semanticLayerService;
+    private SchemaService schemaService;
 
     @Autowired
     private KnowledgeBaseService knowledgeBaseService;
 
     @Autowired
     private SearchMatchStrategy searchMatchStrategy;
-
     @Override
-    public MapInfoResp map(QueryMapReq queryMapReq) {
-
-        QueryReq queryReq = new QueryReq();
-        BeanUtils.copyProperties(queryMapReq, queryReq);
-        List<DataSetResp> dataSets = dataSetService.getDataSets(queryMapReq.getDataSetNames(), queryMapReq.getUser());
-
-        Set<Long> dataSetIds = dataSets.stream().map(SchemaItem::getId).collect(Collectors.toSet());
-        queryReq.setDataSetIds(dataSetIds);
-        MapResp mapResp = chatQueryService.performMapping(queryReq);
-        dataSetIds.retainAll(mapResp.getMapInfo().getDataSetElementMatches().keySet());
-        return convert(mapResp, queryMapReq.getTopN(), dataSetIds);
-    }
-
-    @Override
-    public List<SearchResult> search(QueryReq queryReq) {
+    public List<SearchResult> retrieve(QueryReq queryReq) {
 
         String queryText = queryReq.getQueryText();
         // 1.get meta info
-        SemanticSchema semanticSchemaDb = semanticLayerService.getSemanticSchema();
+        SemanticSchema semanticSchemaDb = schemaService.getSemanticSchema();
         List<SchemaElement> metricsDb = semanticSchemaDb.getMetrics();
         final Map<Long, String> dataSetIdToName = semanticSchemaDb.getDataSetIdToName();
         Map<Long, List<Long>> modelIdToDataSetIds =
@@ -352,133 +320,4 @@ public class RetrieveServiceImpl implements RetrieveService {
         return possibleDataSets.contains(entry.getModel());
     }
 
-    private MapInfoResp convert(MapResp mapResp, Integer topN, Set<Long> dataSetIds) {
-        MapInfoResp mapInfoResp = new MapInfoResp();
-        if (Objects.isNull(mapResp)) {
-            return mapInfoResp;
-        }
-        BeanUtils.copyProperties(mapResp, mapInfoResp);
-        MetaFilter metaFilter = new MetaFilter();
-        metaFilter.setIds(new ArrayList<>(dataSetIds));
-        List<DataSetResp> dataSetList = dataSetService.getDataSetList(metaFilter);
-        Map<Long, DataSetResp> dataSetMap = dataSetList.stream()
-                .collect(Collectors.toMap(DataSetResp::getId, d -> d));
-        mapInfoResp.setDataSetMapInfo(getDataSetInfo(mapResp.getMapInfo(), dataSetMap, topN));
-        mapInfoResp.setTerms(getTerms(mapResp.getMapInfo(), dataSetMap));
-        return mapInfoResp;
-    }
-
-    private Map<String, DataSetMapInfo> getDataSetInfo(SchemaMapInfo mapInfo,
-                                                       Map<Long, DataSetResp> dataSetMap,
-                                                       Integer topN) {
-        Map<String, DataSetMapInfo> map = new HashMap<>();
-        Map<Long, List<SchemaElementMatch>> mapFields = getMapFields(mapInfo, dataSetMap);
-        Map<Long, List<SchemaElementMatch>> topFields = getTopFields(topN, mapInfo, dataSetMap);
-        for (Long dataSetId : mapInfo.getDataSetElementMatches().keySet()) {
-            DataSetResp dataSetResp = dataSetMap.get(dataSetId);
-            if (dataSetResp == null) {
-                continue;
-            }
-            if (CollectionUtils.isEmpty(mapFields.get(dataSetId))) {
-                continue;
-            }
-            DataSetMapInfo dataSetMapInfo = new DataSetMapInfo();
-            dataSetMapInfo.setMapFields(mapFields.getOrDefault(dataSetId, Lists.newArrayList()));
-            dataSetMapInfo.setTopFields(topFields.getOrDefault(dataSetId, Lists.newArrayList()));
-            dataSetMapInfo.setName(dataSetResp.getName());
-            dataSetMapInfo.setDescription(dataSetResp.getDescription());
-            map.put(dataSetMapInfo.getName(), dataSetMapInfo);
-        }
-        return map;
-    }
-
-    private Map<Long, List<SchemaElementMatch>> getMapFields(SchemaMapInfo mapInfo,
-                                                               Map<Long, DataSetResp> dataSetMap) {
-        Map<Long, List<SchemaElementMatch>> result = new HashMap<>();
-        for (Map.Entry<Long, List<SchemaElementMatch>> entry : mapInfo.getDataSetElementMatches().entrySet()) {
-            List<SchemaElementMatch> values = entry.getValue().stream()
-                    .filter(schemaElementMatch ->
-                            !SchemaElementType.TERM.equals(schemaElementMatch.getElement().getType()))
-                    .collect(Collectors.toList());
-            if (CollectionUtils.isNotEmpty(values) && dataSetMap.containsKey(entry.getKey())) {
-                result.put(entry.getKey(), values);
-            }
-        }
-        return result;
-    }
-
-    private Map<Long, List<SchemaElementMatch>> getTopFields(Integer topN,
-                                                               SchemaMapInfo mapInfo,
-                                                               Map<Long, DataSetResp> dataSetMap) {
-        Map<Long, List<SchemaElementMatch>> result = new HashMap<>();
-        if (0 == topN) {
-            return result;
-        }
-        SemanticSchema semanticSchema = semanticLayerService.getSemanticSchema();
-        for (Map.Entry<Long, List<SchemaElementMatch>> entry : mapInfo.getDataSetElementMatches().entrySet()) {
-            Long dataSetId = entry.getKey();
-            List<SchemaElementMatch> values = entry.getValue();
-            DataSetResp dataSetResp = dataSetMap.get(dataSetId);
-            if (dataSetResp == null || CollectionUtils.isEmpty(values)) {
-                continue;
-            }
-            String dataSetName = dataSetResp.getName();
-            //topN dimensions
-            Set<SchemaElementMatch> dimensions = semanticSchema.getDimensions(dataSetId)
-                    .stream().sorted(Comparator.comparing(SchemaElement::getUseCnt).reversed())
-                    .limit(topN - 1).map(mergeFunction()).collect(Collectors.toSet());
-
-            SchemaElementMatch timeDimensionMatch = getTimeDimension(dataSetId, dataSetName);
-            dimensions.add(timeDimensionMatch);
-
-            //topN metrics
-            Set<SchemaElementMatch> metrics = semanticSchema.getMetrics(dataSetId)
-                    .stream().sorted(Comparator.comparing(SchemaElement::getUseCnt).reversed())
-                    .limit(topN).map(mergeFunction()).collect(Collectors.toSet());
-
-            dimensions.addAll(metrics);
-            result.put(dataSetId, new ArrayList<>(dimensions));
-        }
-        return result;
-    }
-
-    private Map<String, List<SchemaElementMatch>> getTerms(SchemaMapInfo mapInfo,
-                                                           Map<Long, DataSetResp> dataSetNameMap) {
-        Map<String, List<SchemaElementMatch>> termMap = new HashMap<>();
-        Map<Long, List<SchemaElementMatch>> dataSetElementMatches = mapInfo.getDataSetElementMatches();
-        for (Map.Entry<Long, List<SchemaElementMatch>> entry : dataSetElementMatches.entrySet()) {
-            DataSetResp dataSetResp = dataSetNameMap.get(entry.getKey());
-            if (dataSetResp == null) {
-                continue;
-            }
-            List<SchemaElementMatch> terms = entry.getValue().stream().filter(schemaElementMatch
-                            -> SchemaElementType.TERM.equals(schemaElementMatch.getElement().getType()))
-                    .collect(Collectors.toList());
-            termMap.put(dataSetResp.getName(), terms);
-        }
-        return termMap;
-    }
-
-    /***
-     * get time dimension SchemaElementMatch
-     * @param dataSetId
-     * @param dataSetName
-     * @return
-     */
-    private SchemaElementMatch getTimeDimension(Long dataSetId, String dataSetName) {
-        SchemaElement element = SchemaElement.builder().dataSet(dataSetId).dataSetName(dataSetName)
-                .type(SchemaElementType.DIMENSION).bizName(TimeDimensionEnum.DAY.getName()).build();
-
-        SchemaElementMatch timeDimensionMatch = SchemaElementMatch.builder().element(element)
-                .detectWord(TimeDimensionEnum.DAY.getChName()).word(TimeDimensionEnum.DAY.getChName())
-                .similarity(1L).frequency(BaseWordBuilder.DEFAULT_FREQUENCY).build();
-
-        return timeDimensionMatch;
-    }
-
-    private Function<SchemaElement, SchemaElementMatch> mergeFunction() {
-        return schemaElement -> SchemaElementMatch.builder().element(schemaElement)
-                .frequency(BaseWordBuilder.DEFAULT_FREQUENCY).word(schemaElement.getName()).similarity(1)
-                .detectWord(schemaElement.getName()).build();
-    }
 }
