+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/web/service/DownloadService.java
@@ -10,5 +10,7 @@ public interface DownloadService {
     void downloadByStruct(DownloadMetricReq downloadStructReq,
                           User user, HttpServletResponse response) throws Exception;
 
-    void batchDownload(BatchDownloadReq batchDownloadReq, User user, HttpServletResponse response) throws Exception;
+    void batchDownload(BatchDownloadReq batchDownloadReq, User user,
+                       HttpServletResponse response) throws Exception;
+
 }
