+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/web/service/SchemaService.java
@@ -1,16 +1,13 @@
 package com.tencent.supersonic.headless.server.web.service;
 
-import com.github.pagehelper.PageInfo;
 import com.tencent.supersonic.auth.api.authentication.pojo.User;
+import com.tencent.supersonic.common.pojo.ItemDateResp;
 import com.tencent.supersonic.common.pojo.enums.AuthType;
 import com.tencent.supersonic.headless.api.pojo.DataSetSchema;
-import com.tencent.supersonic.headless.api.pojo.request.DataSetFilterReq;
+import com.tencent.supersonic.headless.api.pojo.ItemDateFilter;
+import com.tencent.supersonic.headless.api.pojo.SemanticSchema;
 import com.tencent.supersonic.headless.api.pojo.request.ItemUseReq;
-import com.tencent.supersonic.headless.api.pojo.request.PageDimensionReq;
-import com.tencent.supersonic.headless.api.pojo.request.PageMetricReq;
 import com.tencent.supersonic.headless.api.pojo.request.SchemaFilterReq;
-import com.tencent.supersonic.headless.api.pojo.request.SchemaItemQueryReq;
-import com.tencent.supersonic.headless.api.pojo.response.DataSetSchemaResp;
 import com.tencent.supersonic.headless.api.pojo.response.DimensionResp;
 import com.tencent.supersonic.headless.api.pojo.response.DomainResp;
 import com.tencent.supersonic.headless.api.pojo.response.ItemResp;
@@ -19,33 +16,46 @@ import com.tencent.supersonic.headless.api.pojo.response.MetricResp;
 import com.tencent.supersonic.headless.api.pojo.response.ModelResp;
 import com.tencent.supersonic.headless.api.pojo.response.ModelSchemaResp;
 import com.tencent.supersonic.headless.api.pojo.response.SemanticSchemaResp;
+import com.tencent.supersonic.headless.server.pojo.MetaFilter;
+import com.tencent.supersonic.headless.server.pojo.yaml.DataModelYamlTpl;
+import com.tencent.supersonic.headless.server.pojo.yaml.DimensionYamlTpl;
+import com.tencent.supersonic.headless.server.pojo.yaml.MetricYamlTpl;
 
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.ExecutionException;
 
 public interface SchemaService {
 
-    List<DataSetSchemaResp> fetchDataSetSchema(DataSetFilterReq filter);
-
     DataSetSchema getDataSetSchema(Long dataSetId);
 
-    List<DataSetSchema> getDataSetSchema();
+    SemanticSchema getSemanticSchema();
+
+    SemanticSchemaResp fetchSemanticSchema(SchemaFilterReq schemaFilterReq);
 
     List<ModelSchemaResp> fetchModelSchemaResps(List<Long> modelIds);
 
-    PageInfo<DimensionResp> queryDimension(PageDimensionReq pageDimensionReq, User user);
+    List<DimensionResp> getDimensions(MetaFilter metaFilter);
 
-    PageInfo<MetricResp> queryMetric(PageMetricReq pageMetricReq, User user);
+    DimensionResp getDimension(String bizName, Long modelId);
 
-    List querySchemaItem(SchemaItemQueryReq schemaItemQueryReq);
+    List<MetricResp> getMetrics(MetaFilter metaFilter);
 
     List<DomainResp> getDomainList(User user);
 
     List<ModelResp> getModelList(User user, AuthType authType, Long domainId);
 
-    SemanticSchemaResp fetchSemanticSchema(SchemaFilterReq schemaFilterReq);
+    List<ModelResp> getModelList(List<Long> modelIds);
 
     List<ItemUseResp> getStatInfo(ItemUseReq itemUseReq) throws ExecutionException;
 
     List<ItemResp> getDomainDataSetTree();
+
+    void getSchemaYamlTpl(SemanticSchemaResp semanticSchemaResp,
+                          Map<String, List<DimensionYamlTpl>> dimensionYamlMap,
+                          List<DataModelYamlTpl> dataModelYamlTplList,
+                          List<MetricYamlTpl> metricYamlTplList,
+                          Map<Long, String> modelIdName);
+
+    ItemDateResp getItemDate(ItemDateFilter dimension, ItemDateFilter metric);
 }
