+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/web/service/impl/CollectServiceImpl.java
@@ -24,7 +24,7 @@ public class CollectServiceImpl implements CollectService {
     private CollectMapper collectMapper;
 
     @Override
-    public Boolean createCollectionIndicators(User user, CollectDO collectReq) {
+    public Boolean collect(User user, CollectDO collectReq) {
         CollectDO collect = new CollectDO();
         collect.setType(Strings.isEmpty(collectReq.getType()) ? type : collectReq.getType());
         collect.setUsername(user.getName());
@@ -34,7 +34,7 @@ public class CollectServiceImpl implements CollectService {
     }
 
     @Override
-    public Boolean deleteCollectionIndicators(User user, Long id) {
+    public Boolean unCollect(User user, Long id) {
         QueryWrapper<CollectDO> collectDOQueryWrapper = new QueryWrapper<>();
         collectDOQueryWrapper.lambda().eq(CollectDO::getUsername, user.getName());
         collectDOQueryWrapper.lambda().eq(CollectDO::getId, id);
@@ -44,7 +44,7 @@ public class CollectServiceImpl implements CollectService {
     }
 
     @Override
-    public Boolean deleteCollectionIndicators(User user, CollectDO collectReq) {
+    public Boolean unCollect(User user, CollectDO collectReq) {
         QueryWrapper<CollectDO> collectDOQueryWrapper = new QueryWrapper<>();
         collectDOQueryWrapper.lambda().eq(CollectDO::getUsername, user.getName());
         collectDOQueryWrapper.lambda().eq(CollectDO::getCollectId, collectReq.getCollectId());
@@ -54,7 +54,7 @@ public class CollectServiceImpl implements CollectService {
     }
 
     @Override
-    public List<CollectDO> getCollectList(String username) {
+    public List<CollectDO> getCollectionList(String username) {
         QueryWrapper<CollectDO> queryWrapper = new QueryWrapper<>();
         if (!StringUtils.isEmpty(username)) {
             queryWrapper.lambda().eq(CollectDO::getUsername, username);
@@ -63,7 +63,7 @@ public class CollectServiceImpl implements CollectService {
     }
 
     @Override
-    public List<CollectDO> getCollectList(String username, TypeEnums typeEnums) {
+    public List<CollectDO> getCollectionList(String username, TypeEnums typeEnums) {
         QueryWrapper<CollectDO> queryWrapper = new QueryWrapper<>();
         if (!StringUtils.isEmpty(username)) {
             queryWrapper.lambda().eq(CollectDO::getUsername, username);
